Fix Modules Pausebot et Signalsample
