+++ config.example.yml
@@ -132,7 +132,14 @@ trading_options:
   PERCENT_SIGNAL_BUY: False
 
   # MODS SETTINGS_STRING
+
+  # pausebotmod
+  # X of 15 MA's indicating sell  
   PAUSEBOTMOD_THRESHOLD: 14
+  PAUSEBOTMOD_SYMBOL: ETHUSDT
+
+  # custsignalmod
+  # How many of the 26 indicators to indicate a buy
   SIGNALSAMPLEMOD_THRESHOLD : 18
 
 
@@ -158,7 +165,7 @@ trading_options:
   SORT_LIST_TYPE: price_change
 
   SIGNALLING_MODULES:
-    #- modules.pausebotmod
+    - modules.pausebotmod
     #- modules.custsignalmod
     #- modules.custsignalmod_speed
     #- modules.signalsamplemod
